[c10d][fr] Decouple the core logic of FR with the entry and event type
Pull Request resolved: https://github.com/pytorch/pytorch/pull/152585


We want to make FR generic enough so the first step is to make the FR a template struct so that most of common code logic can be reused. The reason for this is that CudaEvent does not inherit c10::Event and we just want to swap the event part so that for NCCL we use CudaEvent and for the rest of backends, we use c10::event.

cc @H-Huang @awgu @wanchaol @fegin @wz337 @wconstab @d4l3k

@imported-using-ghimport

Differential Revision: [D74262695](https://our.internmc.facebook.com/intern/diff/D74262695/)
ghstack-source-id: 15a8f31471557a0b5497b0ca2013b0aac1247ce7